Move Parser into package
