+++ src/main/java/seedu/binbash/BinBash.java
@@ -2,6 +2,7 @@ package seedu.binbash;
 
 import seedu.binbash.storage.Storage;
 import seedu.binbash.ui.Ui;
+import seedu.binbash.parser.Parser;
 import seedu.binbash.command.Command;
 import seedu.binbash.command.ByeCommand;
 import seedu.binbash.exceptions.BinBashException;
