+++ src/main/java/seedu/binbash/parser/Parser.java
@@ -1,4 +1,4 @@
-package seedu.binbash;
+package seedu.binbash.parser;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
