+++ src/test/java/seedu/binbash/parser/ParserTest.java
@@ -1,4 +1,4 @@
-package seedu.binbash;
+package seedu.binbash.parser;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -9,6 +9,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 import java.time.LocalDate;
 import java.util.ArrayList;
 
+import seedu.binbash.ItemList;
 import seedu.binbash.command.AddCommand;
 import seedu.binbash.command.Command;
 import seedu.binbash.command.DeleteCommand;
